Refactor unit tests - add test_utils.dart file for common unit test tools - refactor SampleResponseData
